api/pyproject.toml
[tool.poetry]
name = "runttare"
version = "0.1.0"
description = "Runttare, perkele!"
authors = ["Wille <villeristimaki@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"RPi.GPIO" = "^0.7.1"
aiohttp = "^3.8.1"
asyncio = "^3.4.3"
fastapi = "^0.78.0"
gunicorn = "^20.1.0"
loguru = "^0.6.0"
SimulRPi = "0.1.0a1"
uvicorn = "^0.17.6"
websockets = "^10.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
